+++ src/lib.rs
@@ -31,6 +31,7 @@ extern crate url;
 #[macro_use]
 extern crate serde_derive;
 extern crate serde;
+#[macro_use]
 extern crate serde_json;
 extern crate tokio;
 extern crate tokio_codec;