+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 This release brings a number of breaking changes, all hopefully considered to be *good* :), and some new process to help track future changes. Some notable changes are listed below. The best source of truth will be the updated rustdocs.
 
+* add `expose` option for ports on container builders [#127](https://github.com/softprops/shiplift/pull/127)
 * removed public {Entity}Builder.new() constructors. Use `{Entity}.builder()` interfaces to construct these instead [#125](https://github.com/softprops/shiplift/pull/125)
 * repo labels are now optional [#102](https://github.com/softprops/shiplift/pull/102)
 * interlacing tty [#101](https://github.com/softprops/shiplift/pull/101)